pyproject.toml
[project]
name = "market_researcher_crew"
version = "0.1.0"
description = "market_researcher_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.12"
dependencies = [
    "crewai==0.28.8",
    "crewai-tools",
    "python-dotenv>=1.0.0",
    "selenium>=4.41.0",
    "setuptools<70",
    "webdriver-manager>=4.0.2",
]

# ... keep the rest of your pyproject.toml the same
[project.scripts]
market_researcher_crew = "market_researcher_crew.main:run"
run_crew = "market_researcher_crew.main:run"
train = "market_researcher_crew.main:train"
replay = "market_researcher_crew.main:replay"
test = "market_researcher_crew.main:test"
run_with_trigger = "market_researcher_crew.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
